Add get_mount_search & get_mount_search_from
The existing get_mount only allows you to locate a mount point
by using the file system mounted point.  Sometimes it's useful
to locate the mount points that may be available for a specified
device node.  With this in mind I added a search enum which allows
the caller to match 1 or more mount points by allowing them to
search by any one of the fields in the mount.

Signed-off-by: Tony Asleson <tasleson@redhat.com>